feat: Add new entities and refactor core systems
- **Refactoring `zombie_base`**:
  - Encapsulated member variables (`m_scene`, `m_reanim`, `m_rng`) by making them private.
  - Introduced protected accessor methods (`scene()`, `reanim_sys()`, `rng_sys()`) for derived classes.
  - Updated all dependent code to use the new accessors.

- **Logic Fixes & Improvements**:
  - Corrected the ice path update logic by using `std::min` to prevent incorrect values.
  - Modernized map/set lookups by replacing `.find() == .end()` with C++20's `.contains()` for better readability.
  - Improved `const`-correctness in range-based for loops with structured bindings.

+++ system/plant/starfruit.cpp
@@ -47,14 +47,14 @@ bool plant_starfruit::has_target(object::plant& p) {
 
         auto predict = zombie_base(scene).predict_after(z, static_cast<float>(t));
 
-        if (px < predict + (zr.width / 2) && predict - (zr.width / 2) < px) {
+        if (px < predict + (zr.width / 2.0) && predict - (zr.width / 2.0) < px) {
             return true;
         }
 
-        int x_predict = static_cast<int>(predict + (zr.width / 2) - px);
-        int y_predict = (zr.y + (zr.height / 2) - py);
+        int x_predict = static_cast<int>(predict + (zr.width / 2.0) - px);
+        int y_predict = static_cast<int>(zr.y + (zr.height / 2.0) - py);
 
-        auto deg = atan2(y_predict, x_predict) * 57.2957763671875;
+        auto deg = atan2(static_cast<double>(y_predict), static_cast<double>(x_predict)) * 57.2957763671875;
 
         if (abs(static_cast<int>(p.row) - static_cast<int>(z.row)) >= 2) {
             if (deg > 25 && deg < 35 || deg > -38 && deg < -28) {
